Rename Fetcher to Lookup, Twitter logo adjust
